changed recipe hash format and refactored
